remove prefetch feature in inclusive MSHR.scala
